Add files via upload
The   Bank management systemis an application for maintaining a person’s account  in  a  bank.  The  system  provides  the  access  to  the customer  to  create  an  account,  deposit/withdraw  the  cash  from his  account,  also  to  view  reports  of all  accounts  present.  The followingpresentation provides the specification for the system.The objective is to create application software which can manage all about the customers currently working in bank in order relative them from their manual accounting system.